requirements.txt
reportlab
svglib
Pillow
cairosvg
